+++ config/settings.php
@@ -70,6 +70,7 @@ return [
     'defaults' => [
         'logo' => env('DEFAULT_LOGO', '/images/logo.svg'),
         'room_limit' => env('DEFAULT_ROOM_LIMIT',-1),
+        'own_rooms_pagination_page_size' => env('OWN_ROOMS_PAGINATION_PAGE_SIZE',5),
         'pagination_page_size' => env('DEFAULT_PAGINATION_PAGE_SIZE', 15),
     ]
 ];