+++ routes/api.php
@@ -78,6 +78,8 @@ Route::prefix('v1')->namespace('api\v1')->name('api.v1.')->group(function () {
         });
 
         Route::get('users/search','UserController@search')->name('users.search');
+
+        Route::get('roomTypes', 'RoomTypeController@index')->name('roomTypes.index');
     });
 
     Route::middleware('can:view,room')->group(function () {
@@ -86,6 +88,10 @@ Route::prefix('v1')->namespace('api\v1')->name('api.v1.')->group(function () {
         Route::get('rooms/{room}/join', 'RoomController@join')->name('rooms.join')->middleware('room.authenticate');
         Route::get('rooms/{room}/files/{file}', 'RoomFileController@show')->name('rooms.files.show')->middleware(['can:downloadFile,room,file', 'room.authenticate']);
     });
+
+
+
+
 
     Route::get('meetings/{meeting}/endCallback','MeetingController@endMeetingCallback')->name('meetings.endcallback');